Use size_t

+++ include/sorts.h
@@ -1,11 +1,11 @@
 #ifndef SORTING_ALGORITHMS_SORTS_H
 #define SORTING_ALGORITHMS_SORTS_H
 
-void bubble_sort(int array[], int size);
-void insertion_sort(int array[], int size);
-void selection_sort(int array[], int size);
-void merge_sort(int array[], int size);
-void quick_sort(int array[], int size);
-void counting_sort(int array[], int size);
+void bubble_sort(int array[], size_t size);
+void insertion_sort(int array[], size_t size);
+void selection_sort(int array[], size_t size);
+void merge_sort(int array[], size_t size);
+void quick_sort(int array[], size_t size);
+void counting_sort(int array[], size_t size);
 
 #endif
