+++ src/sorts/selection_sort.c
@@ -1,6 +1,7 @@
+#include <stdio.h>
 #include "utils.h"
 
-void selection_sort(int array[], int size) {
+void selection_sort(int array[], size_t size) {
     for (int i = 0; i < size; ++i) {
         int index = i;
         for (int j = i + 1; j < size; ++j) {
